More clean up: no more SocketHolder, more meaningful Connection
... and move EIEIOMessageHandler to better place
